requirements.txt
PyMuPDF
Pillow
pytesseract
tqdm
python-dotenv
langchain
python-docx
openai
streamlit
langchain_community